backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
firebase-admin==6.5.0
google-generativeai==0.8.3
httpx==0.27.2
python-multipart==0.0.12
